backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
pydantic==2.11.7
pydantic-settings==2.10.1
python-multipart==0.0.20
celery==5.5.3
redis==6.4.0
requests==2.32.5
python-pptx==1.0.2
pypdf==5.9.0
python-docx==1.2.0
beautifulsoup4==4.13.4
duckduckgo-search==7.5.5
openai==1.102.0
anthropic==0.64.0
